% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuikR-package.R
\docType{package}
\name{tuikr-package}
\alias{tuikr}
\alias{tuikr-package}
\title{tuikr: Access Statistical and Geographic Data from TUIK}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Accesses Turkish Statistical Institute (TUIK) statistical tables, portal resources, SDMX dataflows, and geographic data through the public TUIK web portals.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://eremrah.com/tuikr/}
  \item \url{https://github.com/emraher/tuikr}
  \item Report bugs at \url{https://github.com/emraher/tuikr/issues}
}

}
\author{
\strong{Maintainer}: Emrah Er \email{eer@eremrah.com}

}
\keyword{internal}
